daemon,greeter: Make greeter/autologin session initiate PAM conversation
Now the greeter (and also the autologin code) has to say what
PAM stack it wants the slave to run.  When that stack is ready,
we emit the Ready signal as before, but now the Ready signal
carries a string argument saying which service is ready to
converse.

When we support multiple PAM stacks, the greeter will call
StartConversation for each stack, and will keep the UI
associated with each stack disabled until the Ready signals
come back one-by-one.