greeter: Add a login extension mechanism to greeter
This allows extensions to drive which PAM conversations
get run and potentially augment the login window UI.
This commit adds one builtin extension that
does the traditional unified authentication thing, and
a plugin for password-only based authentication.

By default we use the builtin extension, but enable
the plugin with --enable-split-authentication

Subsequent commits will add support for fingerprint and smartcard
plugins.