Add block-console-session-requests display property
This property is just a hint for the display factory to
queue SessionToAdd requests from consolekit instead of
processing them immediately.

This will be useful for the slave to lock a display during
the critical moment after the login window session is closed
before the user's session is opened.

Note, the display factory doesn't look at the hint yet, and the
slave doesn't set it, yet, either.