Fix Other... user visibility check
Before we'd only add the Other item if the number of users
was 1, when we should have been adding it if the number of
users is greater than or equal to 1.  This is because in some
cases more than one item can get added to the user list before
the visibility function is run.