Fix ellipses usage on User Switch Applet
"Account Information" and "System Preferences" have ellipses "..." on the User
Switch Applet. Those are normally meant to indicate that part of the action
needs further user input. Both of these just open some app, so IMHO the
ellipses are not needed here: the action is complete when the app is started
(Edit->Preferences in GNOME apps does not have ellipses for this same reason)

The HIG says:

---
Label the menu item with a trailing ellipsis ("...") only if the command
requires further input from the user before it can be performed. Do not add an
ellipsis to items that only present a confirmation dialog (such as Delete), or
that do not require further input (such as Properties, Preferences or About).
---

See: https://bugzilla.gnome.org/show_bug.cgi?id=568866