Do not overwrite $LANGUAGE and other locale variables
gdm does not have a facility to specify a list of languages for $LANGUAGE, so
do not break $LANGUAGE by forcing it to $LANG (they have a completely different
syntax, and different meanings). Also, if the system sets $LC_ALL (which
distros should never ever do), we should not clobber this.

$LINGUAS is not a runtime environment variable for locales, so just drop this.

Bug: https://bugzilla.gnome.org/show_bug.cgi?id=603853
Bug-Ubuntu: https://launchpad.net/bugs/407300