Ensure users-loaded signal is emitted
The users-loaded signal is used to indentify the point when it is acceptable
to call gdm_user_manager_list_users().  This signal normally
gets emitted after getting a list of frequently logged in users from ConsoleKit.
In the event the ConsoleKit daemon isn't available, however, the signal was
never getting emitted.  This commit causes the signal to get emitted from those
cases as well.

https://bugzilla.gnome.org/show_bug.cgi?id=620908