Revert "Update duplicate real names to be unique"
This reverts commit dfc66ca1f169c8b9f1c504f762b3d01605917a98.
From https://bugzilla.gnome.org/show_bug.cgi?id=549349

This is a very expensive change and one that I think can be done
in a better way.  It doesn't make sense to me to have automatically
generated and changing display names.  We want the user's display name
to be editable and discoverable using the user account settings tool.

There are already a few way to disambiguate the user identity.  The first
is the name field.  It can be changed to add a "(work)" label or whatever
the user wants.  Another is the account icon.  And lastly the tooltip over
the user list entry will show the username for the account.

https://bugzilla.gnome.org/show_bug.cgi?id=620908