Don't exit when local display fails to start
It doesn't really make a lot of sense to exit when the
local display can't be started.  Exiting won't leave the system
in any better shape that it would be with gdm running in the
background.  On the other hand it makes it so that XDMCP and other
services can't run headless.  Ideally we'd have some kind of
smart detection of when local hardware is available and when it
changes.  And smarter fallback code for when the hardware is
available but the display can't be started due to some error.

https://bugzilla.gnome.org/show_bug.cgi?id=342397