Don't load users in user manager automatically
In theory, the user manager could be useful for some use cases
without loading the entire user list.

In practice, the user manager requires the list to be loaded
before it's really functional.

Still, that could potentially be fixed in the future.

commit 9ab0c955326c9be530757dde8f31594b88593c69 mentions making
this change, already, but actually neglects to do it.

This commit just "fixes the glitch".

https://bugzilla.gnome.org/show_bug.cgi?id=622639