Fix up translations after string freeze break
Commit 1b33435081074746ba3408def87b3f119ce919b6
fixed a problem with GDM in low resolution machines.

While this commit is a valid and important bug fix,
it unfortunately, broke string freze.

See:

http://mail.gnome.org/archives/gdm-list/2010-June/msg00014.html

This commit manually fixes up all the msgids and msgstrs so that
the old transitions still work.

Note, some of the translations get plurality wrong for "Session"
(like the original C string did before commit 1b33435).  That's
not optimal, but not any worse than before.