Always show “other…” choice
When the system is configured to use NIS or LDAP, with no local
user, no user shows up in the chooser. In this case, there is
specific code that makes the “other…” choice disappear. I can’t
think of any case where it would be useful, and in this case it
makes logging in entirely impossible.

https://bugzilla.gnome.org/show_bug.cgi?id=624467