Don’t use --exit-with-session for greeter dbus-launch call
The session bus daemon for the login session is started with
--exit-with-session. This starts a babysitter process that
watches for the loss of X connectivity or SIGHUP. However, when
the login session is terminated, the X server is not reset and
SIGHUP is not sent. This leads to leaking the babysitter
process, which remains here until the user logs out.

Since the session bus daemon is killed by hand, there is no need
for --exit-with-session.