greeter: frob session defaults in gsettings world
This commit changes the defaults for various parts of the
GNOME,, so that things are properly locked down again
in the greeter session.

The current gconf overrides don't apply for recent
components, since they have been ported to gsettings.

The approach used here requires a binary gsettings database
to get generated during "make dist" time and shipped in the
tarball.  This is suboptimal.

In future versions of dconf we'll be able to ship keyfiles
instead and then generate the database in package %post.