daemon: set sigpipe to default value before starting session
Some users have complained that because GDM sets SIGPIPE to ignore that
this causes problems when running programs started in the user session.
For example, if the user runs "cat file | head -10" the head command
completes after processing 10 lines and sends a SIGPIPE to cat to inform
it to exit.  However, when SIGPIPE is set to ignore, the CPU is wasted
while cat completes.

This commit fixes this issue and sets SIGPIPE to SIG_DFL
before running the Xsession script.  Note this is done after fork() so
it doesn't affect the gdm-session-worker process.

https://bugzilla.gnome.org/show_bug.cgi?id=630484