greeter: don't tank if accounts service removes user we don't know about
The chooser right now will throw a critical if we try to remove an item
that doesn't exist.

There are some situations where users get removed from the account
service that we never added to the user list.  In those cases, we
need to be careful not to try to remove them to prevent crash-inducing
criticals.

https://bugzilla.gnome.org/show_bug.cgi?id=644149