greeter: disable the login button in MODE_AUTHENTICATION if no query is pending
If the user does not need a password to login with gdm, then hitting the
"log in" button causes gdm to crash. This seems to be due to an
AnswerQuery DBUS message being generated even if the password input
field is not visible and an password has not been requested. The user
has a real opportunity to hit this bug at least if pam_mkhomedir is in
use and needs some time to finish due to the size of the skeleton.

This commit desensitizes the Login button when it's unavailable.

https://bugzilla.gnome.org/show_bug.cgi?id=643176