simple-greeter: Don't request an invalid size
GTK widgets must at all times report a size they can handle. So it is
not allowed to return 0 when not realized, because then size allocations
break when GTK uses this size for its widget.
In this case, GTK uses the pre-realize size to determine the size it
should request when creating the greeter window - chicken and egg so to
say.

This patch just uses the default monitor (I guess the root window's
monitor is the default monitor?) for determining the login window size.

One thing this patch doesn't do is add a call to gtk_widget_queue_resize()
from the realize callback or from monitor-changing signals, though
that's probably technically necessary.

https://bugzilla.gnome.org/show_bug.cgi?id=646498