greeter: force symbolic icons in panel
gtk falls back to non-symbolic icons before looking
for symbolic icons in parent icon themes.

For the greeeter panel, though, any symbolic icon is better than a
non-symblic icon, so that behavior is wrong.

This commit changes gdm to not check for fallbacks in the absense of
a symblic icon. The icons we care about are shipped in the base
icon theme anyway, so we can depend on them being available.
(cherry picked from commit ae0b1d334956fbb7b0a677a462aefdf9c46ecde4)