daemon: clean up greeter when unused
When user switching GDM needs to show a transient greeter for
the user to pick a name from the list and jump to that session.

After the user has been jumped to the selected session we don't
really have a use for the greeter session.  Historically, we've
kept it around anyway, though, because X by default jumps back
to the VT it started on when it exits, and we don't want the
user to get thrown to an empty VT after they log out.

This commit changes X to get started with the "-novtswitch"
option, so that it doesn't do the undesirable switch-on-exit
thing.

This allows us to clean up the useless greeter following user
switches.

Based on work by Josselin Mouette <joss@debian.org>

https://bugzilla.gnome.org/show_bug.cgi?id=618047