daemon: parameterize create_display
There's some logic in create_display and friends,
that would be good to reuse for transient displays.

This commit adds a type argument to create_display
to say what kind of display to create.

Right now it only supports the one type it's always
supported, but that will change in a follow up commit.

https://bugzilla.gnome.org/show_bug.cgi?id=618047