daemon: share display respawn logic betewen static and transient displays
Right now we ignore transient displays after they're created.  We make
no attempt to clean them up from the display store or put the user on
a login screen when their display ends or anything like that.

This commit changes the display factory to monitor transient display
status in much the same way it monitors static display status.

Note for static displays we always respawn when finishing, where
as with transient displays either die quietly, jump to an existing
login display, or respawn.

https://bugzilla.gnome.org/show_bug.cgi?id=618047