daemon,greeter: login after PAM has had time to talk to user
The greeter currently gives the go ahead for the session to start
as soon as the user has been authorized.  Then the slave quickly
runs through the remaining hoops and the greeter gets promptly
killed.

This commit changes the logic so that the slave doesn't get the
go ahead to start the session until all the hoops have been
run through, and potential messages have been queued, and displayed.

https://bugzilla.gnome.org/show_bug.cgi?id=583856