daemon: add switch-on-finish property to display
When a display exits, sometimes it's desirable to
jump to a login screen.

This depends if the display is transient or not, and
also whether the display is hosting an already logged
in session or not.

This commit adds a property to the display object that
says whether a switch should happen or not.

Note the code to actually perform the switch will get
added in a follow up commit.

Based on work by Josselin Mouette <joss@debian.org>

https://bugzilla.gnome.org/show_bug.cgi?id=618047