Allow .xsession-errors to be a symlink or FIFO
Don't rename .xsession-errors to .xsession-errors.old if the file is not a
regular file.  The later code will truncate the file to zero anyway, so even in
that case the file won't grow indefinitely.

This is handy when the home directory is on NFS and you want to avoid network
contention when there are buggy programs which spam .xsession-errors.

Also drop the second check that .xsession-errors is a regular file even if the
opening succeeded. With that, we'd open a temporary .xsession-errors.XXXXXX
even if ~/.xsession-errors was perfectly writable (such as being a symlink to
/dev/null).

Bug: https://bugzilla.gnome.org/show_bug.cgi?id=639527
Bug-Ubuntu: https://launchpad.net/bugs/771661