daemon: use gnome-session session files instead of autostart
Before we were doing some sort of weird hybrid thing with
a session file and an autostart directory that wasn't that
much different than just having an autostart directory by
itself.

Now we fully define the session component list from the session
file, and merely provide a pool of new candidate desktop files to
select that sessoin from.

This modernizes how we use gnome-session and as a side-effect
enables us the ability to have fallback sessions (which will
be important when defaulting to a shell based greeter later).