common: don't tank on criticals
Tiny (or not so tiny) bugs in components below GDM in the stack
can lead to critical warnings.

In development releases, GDM by default will crash when encountering
a critical warning.

Losing the login screen is a big deal.  We shouldn't automatically subject
our users to "can't log in" in order to highlight bugs in development
releases.

This commit removes fatal criticals by default.