slave: pass seat id from product/simple slave to worker, via the direct session
Pass seat id from product and simple slave to the direct sessions
created by it (along with the other display meta data we pass already).

Pass seat id from session to worker (via D-Bus) along with the other
display meta data.

We need the seat information for the PAM conversation (later patch)
hence we need to pass it from the slaves down to the workers.

(Note that the seat ID has always been passed from the display to the
slave, we just need to pass it on the worker now, so that the chain is
complete: display → slave → session → worker)