Lock down log out in the login session
It never makes sense to offer log out when you are not logged in,
and setting this lockdown key is an easy way to make gnome-sessions
'fail whale' dialog more appropriate for the login session. With
this key set, it will just say

A problem has occurred and the system can't recover.
Please contact a system administrator

https://bugzilla.gnome.org/show_bug.cgi?id=658552