daemon: stop greeter session not greeter worker
Since changing the greeter to run in its own PAM
session, we've started killing the worker process
instead of the session process right before the
user's session is started.

This is wrong, because it means we don't give the
greeter session time to kill itself in an orderly
way before we rip away its session bus.

This commit changes the welcome session stop function
to kill the pid of the welcome session instead of
the pid of its worker parent. This change requires
reinstrumenting all callers of gdm_welcome_session_stop
to not free the session until the "stopped" signal
following gdm_welcome_session_stop.

https://bugzilla.gnome.org/show_bug.cgi?id=658481