daemon: set initial keyboard layout before starting server
If we don't do this magic incantation then the keyboard
layout selector won't work until the user hits some key on their
keyboard.

Since we have to do this incantation anyway, this commit also gives
preferential treatment to the "us" keyboard layout.  This makes sense
to do because most usernames and passwords are ascii.