daemon: add better unified authentication compatibility
The daemon and fallback greeter support --disable-split-authentication,
but the new shiny greeter doesn't.  This posed a problem for
distributions that want to use the new shiny greeter (in an albeit
degraded mode) but don't yet have support at the system level for
split authentication.

This commit adds a small amount of goo to make the split authentication
pam service names translate to the unified pam service name and
thus give the shell greeter a level of compatibility with
--disable-split-authentication.