data: drop weird dconf script
It adds a lot of packaging pain and confusion for distributors,
and the bulk of the gain from it could be fixed by adding a
setting to gnome-settings-daemon.

Instead, just ship a static list of settings.

Distros will still need to run dconf update in their post transaction
hooks, however.