daemon: Don't emit session-exited when non-authenticated worker fails
Sometimes PAM modules are finicky and don't die when you tell them to.
Instead they fail some seconds later.

If a user successfully logs in with one stack and another stack is
being troublesome, then we'll get notified about it finishing up
after the user is already logged in.

When that happens, we erroneously assume the stack finishing is the
stack the user's session is running on and then proceed to log the
user out.

This commit makes us be a little more careful about our bookkeeping
so we can ignore failures from slow PAM modules.