daemon: don't every SIGKILL children
There's no situation where it's the right thing to do.
The children of the daemon are responsible for themselves.
If they don't go away when asked we shouldn't second guess
them, we just need to ignore them.

This means we may end up with zombie children if those children
have bugs, but it's better than prematurely killing them if they're
slow.