session: add username arg to SetupForProgram
Previously, SetupForProgram was only used to start a greeter,
which always runs as the gdm user.

Now SetupForProgram is potentially used to start the initial setup
assistant.  This program can run as a different user.

This commit adds a "username" arg to SetupForProgram that defaults
to "gdm" for the greeter, and becomes "gdm-initial-setup" for the
initial setup assistant.