welcome: Use Linux PR_SET_PDEATHSIG to ensure dbus-daemon dies when we do
It's convenient for debugging to "kill -9 $(pidof gdm-binary)", but
presently because gdm runs bare "dbus-launch" which daemonizes,
it won't terminate when the parent gdm welcome session dies.

Other approaches here are to use "dbus-daemon --no-fork --print-address"
directly as a child, and kill it when we get SIGTERM:
http://git.gnome.org/browse/at-spi2-core/tree/bus/at-spi-bus-launcher.c

But this is an easy fix, even if it only works on Linux.

https://bugzilla.gnome.org/show_bug.cgi?id=674982