Add --with-default-pam-config option, autodetect from /etc/foo-release files
The PAM files that ship with GDM are really specific to Red Hat's
historical fork of pam.  For example, the "system-auth" file still
lives in the Fedora 17 "pam" git.  A long while back, Debian hit the
same problem, and of course the difference is the naming;
common-auth/common-password etc.

OpenEmbedded then picked up Debian's PAM fork.  Since for OSTree-GNOME
we're using Poky/OpenEmbedded, let's add an option to integrate with
their PAM.

We use code similar to what NetworkManager has, so we should keep
using the Red Hat files on systems with /etc/redhat-release or
/etc/fedora-release.

https://bugzilla.gnome.org/show_bug.cgi?id=675085