daemon: don't freak out if greeter dies after login finishes
If we've already logged a user in and the greeter is shutting
down and crashes, we should just ignore the crash and continue
with the login process.

Furthermore, if we explicitly kill the greeter and it dies with
a TERM signal, then we definitely don't want to treat that as
if it was a crash.