daemon: Port GdmSession to use GDBus
GdmSession is an object in the slave that manages
the various session worker processes.  Each session
worker process talks to PAM to perform authentication
for the user. For instance, if the user has a fingerprint
reader, then there will normally be two worker processes,
one for handling fingerprint auth, and one for handling
password auth.  GdmSession is the interface layer in the
slave to talking to those running worker processes.

This commit ports GdmSession over to GDBus from dbus-glib.

https://bugzilla.gnome.org/show_bug.cgi?id=622888