daemon: drop ck connector
ck-connector is copy-and-paste code used in various parts of
the freedesktop stack for interfacing with consolekit.

It's legacy code, that uses dbus-glib.  We want to drop our
dbus-glib dependency, so this is one obstacle in the way.

One option would be to port it to gdbus, but that would
require updating all copies everywhere (or run into potential
conflicts when, say, pam_ck_connector's version of the
ck-conector code ends up clashing with our updated version).

This commit, instead, just drops ck-connector and uses direct
gdbus calls instead.  We're only depending on ck-connector for
two calls in a legacy path anyway, so it's actually a net
reduction in code.

https://bugzilla.gnome.org/show_bug.cgi?id=622888