daemon: Port GdmManager to GDBus
The GdmManager object controls the GdmLocalDisplayFactory and
GdmXdmcpDisplayFactory singleton objects, which manage displays on local
VTs and displays on remote machines respectively.

Another role of the GdmManager object is to aggregate and export the
displays currently being managed by those display factories over the
system bus.

This commit moves GdmManager over to using GDBus and the
GDBusObjectManager interface for display enumeration.

https://bugzilla.gnome.org/show_bug.cgi?id=622888