daemon: Add an interface for communicating with GDM via D-Bus
One goal for GNOME 3.6, is to replace the screen locking functionality
provided by gnome-screensaver with redesigned functionality provided
by gnome-shell.

At the same time, it makes sense to consolidate the yucky PAM
authentication code to one place (GDM).

Right now only greeters can talk to GDM.  At the time the greeter is
started, the slave sets up a private communication channel which the
greeter then connects to for initiating communication.

This commit adds a new method to the org.gnome.DisplayManager.Manager
interface that allows opening a private connection to the slave that
is associated with the currently running session.  That slave exports
the session object over the bus that greeters can interact with the
session as appropriate.  This interface replaces the
GDM_GREETER_DBUS_ADDRESS environment variable that used to to be used
for connecting the greeter to the slave.

This commit also drops gdm-greeter-server and gdm-chooser-server which
don't fit the new model, and are really just thin middle men that don't do
anything important.

Furthermore, this commit splits GdmSession interfaces 3 orthogonal parts
up into 3 separate interfaces on the session object.

A future commit will make this interface work for screensavers/reauthentication.

Based on work by Giovanni Campagna <gcampagna@src.gnome.org>

https://bugzilla.gnome.org/show_bug.cgi?id=676381