libgdmgreeter: rename to libgdm
libgdmgreeter is useful for clients that aren't greeters.
This will be even more true in the future as it gets updated
to be useful for screensavers.

Given it offers APIs that apply to non-greeter applications,
it's silly to have greeter in the library name.

This commit renames libgdmgreeter to libgdm.

https://bugzilla.gnome.org/show_bug.cgi?id=676381