daemon: clean house (drop factory mode)
GDM has a currently unused feature called "factory mode", where
the login screen gets allocated its own VT and any subsequent logins
from that login screen go to their own VT.  Any time there's a user
switch GDM just jumps back to initial VT where the login screen is
patiently waiting.

This feature has a lot of upsides, and we've had it as a TODO item to
start using it for while now.

It doesn't look it's going to happen in the near term, though, and
factory mode has downsides as well:

- a VT switch after the user hits enter at the login screen would be
  jarring and would ruin the "flicker free" boot experience we've tried
  to acheive at various points in the past. This could theoretically be
  fixed by Wayland.

- it adds a bunch of untested, unused code to the codebase.

This latter reason makes me want to kill it for now. It shouldn't be
hard to resurrect later if we end up needing the feature.

This commit drops that code.