daemon: Port GdmChooserServer to GDBus
GdmChooserServer is the slave-side object that handles communication
with choosers. The chooser talks over a private peer-to-peer dbus
connection to its slave via GdmChooserServer.

A chooser is like a greeter, but instead of presenting a login screen
it presents a list of hosts on the network that will offer login screens
if asked.  The user picks one of these hosts and their display is
redirected to it.

This commit makes GdmChooserServer use GDBus.

This gets us one step closer to GDM running without dbus-glib.

https://bugzilla.gnome.org/show_bug.cgi?id=6228