daemon: Drop manager bus watching code
GdmManager watches the system bus to see whenever a service
leaves.  The intent of the code is to stop managing displays
associated with slaves that have disappeared.

The code wasn't ever finished, though, so drop it for now.

It will be easy to add back in a different way after GdmManager
has been ported to GDBus.