daemon: NUL terminate X11 cookie
The X11 cookie is used as a shared secret between X clients
and the X server.  This cookie is written into an Xauthority
file pointed to by the XAUTHORITY environment variable.  In the
event standard peer-credential based access isn't available, clients
read this file and gain authorization to use the X server by
presenting the cookie to the X server.

This cookie is passed from the gdm daemon to the slave over the bus
as an array of bytes. GDBus expects bytes strings to be NUL terminated.

This commit NUL terminates the cookie, so that it will be compatible
with GDBus's generated code.