chooser: port to libgdm
The chooser is the UI showns to XDMCP clients connecting
via indirect queries.  It shows a list of login screens
on the network and lets the user pick that login screen
to jump to.

Right now it uses its own hardcoded D-Bus calls
to interact with the GDM daemon.

The D-Bus APIs have changed, though, and so now it
no longer works.

This commit changes simple-chooser to use libgdm
instead.

https://bugzilla.gnome.org/show_bug.cgi?id=622888