simple-greeter: only hide Other item if we can skip it
Normally if "Other" is the only thing we'll show, we forgo it
and jump straight to the password prompt.

There is a bug, though, if multiple authentication methods are
enabled.  In that case, we hide the user chooser, but don't jump
to the Password prompt.

This commit makes sure we show the user chooser / Other in that case.