daemon: Add initial setup integration
When a system boots for the first time, we want to show a special
tool that will allow the user to set up their system the way they
want to. This will be triggered by a special file:

    /var/run/wants-initial-setup

The responsibilities of this tool include creating the user's
account, so we have to create a special user account to run the
tool under.

Administrators are given the ability to turn this off in a GDM
setting if they want to.

https://bugzilla.gnome.org/show_bug.cgi?id=678057