daemon: Replace old method/signal-based API with async method calls
Before, the session worker and session communicated by a series of signals
and methods... but backwards. The session worker would listen for a series
of signals sent out by the session, and respond back by calling methods on
it. This requires a lot of annoying, silly manual labor when trying to add
another method to the API.

So, reverse the API so that the worker manager calls async methods on the
worker itself.

https://bugzilla.gnome.org/show_bug.cgi?id=678057