daemon: don't kill remote session slave at login
When a user disconnects from an XDMCP session we need
to clean it up the associated slave.

The function that did that was also erroneously cleaning up
the slave when the greeter disconnected following the user
successfully typing their password.

This commit guards against the latter problem by checking if
the session is started before deciding whether or not to
kill the slave.

https://bugzilla.gnome.org/show_bug.cgi?id=680314