daemon: hack around timed login not working
The greeter waits for the TimedLoginRequested signal to know
when to initiated the timed login process.

The slave used to send this signal shortly after the greeter
called StartConversation.  Now the greeter doesn't need to
call StartConversation, since conversations are started
implicitly.  This means the slave doesn't have a good place
to know when to emit the TimedLoginRequested signal.

For now, emit the signal a couple of seconds after the greeter
connects. This is a racy stop-gap. We're going to need to redo
how timed login is orchestrated.

https://bugzilla.gnome.org/show_bug.cgi?id=680348