slave: rework stop code paths
Previously code would emit stopped signal directly, and rely on
the side effect of main dropping the last reference to do the actual stop
(including killing X). As this would break in case of memory leaks,
move to calling stop when the session exits.
Also, ensure that stop can be called multiple times (as it still called
from finalize).

https://bugzilla.gnome.org/show_bug.cgi?id=682571