slave: fix memory leak in gdm_simple_slave_open_reauthentication_channel
The initial reference to the GSimpleAsyncResult was never cleared, as
removing from the hash table only cleared the added one.
Also, move removing from the hash table to GdmSession signal handler,
as shown in GIO documentation, as g_simple_async_result_complete_in_idle
ensures that the result stays alive. This way we can use the async function
as the source_tag.

https://bugzilla.gnome.org/show_bug.cgi?id=682571