daemon: reset exec context after fork()
When pam_open_session finishes, the session worker
is set up such that the next fork()/exec() may transition the
user to a user specific context (such as staff_t).

This makes sense for the first fork()/exec() (which is the user
login), but the worker may fork()/exec() other workers after login
for unlock operations.  These workers need to run in a gdm context
not a user context.

This commit changes gdm-session-worker to manually reset the exec()
context after the first fork().

https://bugzilla.gnome.org/show_bug.cgi?id=683426