slave: fix autologin
An inverted condition in the wants_autologin function means we tried
to start greeter when autologin is runnings.

Since autologin is usually much faster than loading GL/gnome-shell, this bug
is sort of spotty to reproduce.

This commit changes the "delay > 0" check to "delay == 0" which is a
correct indicator of autologin.

https://bugzilla.gnome.org/show_bug.cgi?id=682465