daemon: Major cleanup of greeter environment setup
First, we can't call a lot of this stuff inside a
GSpawnChildSetupFunc; for example, g_warning() can try to lock a
mutex; if another thread happened to be holding that mutex from before
we forked, we'd deadlock.  Thus, it needed to be extracted.

Second, just drop the group-name property; nothing was using it, and
it complicated the code.

Third, the error handling was totally inconsistent and ugly; sometimes
we would g_warning, other times we'd re-throw to the caller, other
times we'd do both.  Clean this up by consistently propagating errors
up until the first public API that doesn't take a GError.

https://bugzilla.gnome.org/show_bug.cgi?id=630485