worker: Copy environment from login session to reauth sessions
Reauthentication sessions depend on having the same environment
as the session they were initiated from.  This is important to
make sure login prompts are in the right language, to make sure
the kerberos credentials cache is looked up, and for various
other reasons.

This commit copies the environment from the login session to
any new reauthentication sessions that get started after login.

https://bugzilla.gnome.org/show_bug.cgi?id=684241